requirements.txt
alpaca-py==0.34.0
annotated-types==0.7.0
beautifulsoup4==4.12.3
boto3==1.35.81
botocore==1.35.81
certifi==2024.12.14
charset-normalizer==3.4.0
frozendict==2.4.6
html5lib==1.1
idna==3.10
jmespath==1.0.1
lxml==5.3.0
msgpack==1.1.0
multitasking==0.0.11
numpy==2.2.0
pandas==2.2.3
peewee==3.17.8
platformdirs==4.3.6
pydantic==2.10.3
pydantic_core==2.27.1
python-dateutil==2.9.0.post0
pytz==2024.2
requests==2.32.3
s3transfer==0.10.4
scipy==1.14.1
six==1.17.0
soupsieve==2.6
sseclient-py==1.8.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
webencodings==0.5.1
websockets==14.1
yfinance==0.2.50
